Work around Yahoo!/AOL DMARC changes.
If the sender is using an AOL or Yahoo! email address, put our bounce
handling address in the From header, and the sender in the Reply-To
header. This is because Yahoo! and AOL have both switched their DMARC
settings to p=reject, meaning we are unable to send any email on behalf
of Yahoo!/AOL users to anyone who checks and honours DMARC settings.

Update commonlib to get the test harness fix.