Match up MP for year of stats correctly from postcode
Previously we were looking for the stats for the current MP who may
not have been MP in previous years. This was returning an error. We
now fetch all the reps for an area and then loop over them till we
find one who is both in that years results and has a matching
constituency name.

Fixes #247