use HTML for confirmation emails
Generate an HTML Email for the confirm message and sent it as part of
a multipart/alternative email.

The header and footer are separate templates that are string concatenated
with the main body of the email rather than moving to a full template
system.

Common CSS variables are stored in a module which are then passed in to
the template as params.

The reason for the style params is because you need inline styles for
portable HTML emails becuase HTML email and we don't want to have to
repeat these across all the templates so...

This does not attach the logo image.