Add a donate banner for Christmas and New Year 2016
This is a cherry-pick of commit b8b580c4ddba973c3bc by Matthew
as a basis for the banner for 2016.

I've added utm_* GET query parameters for the donate banner for Google
Analytics. These should help to track the effectiveness of having this
banner in place over the Christmas and New Year period.