Remove link to stats from questionnaire emails (#441)
* Remove link to stats page

We've taken a conscious decision not to update the responsiveness stats in the past few years. They are now at least four years and three general elections out of date and it no longer makes sense to refer questionnaire respondents there.

* Remove link to stats page

Same change for `questionnaire.html` as for `questionnaire` template.