Extend "a few minutes" wait time
It's common for people to email WriteToThem support saying their confirmation email hasn't arrived, and when you check in the database, the message is already confirmed and sent - i.e. they confirmed themselves, having just not waited long enough for the email to come through. If we tell them a longer potential wait time (and use more precise words than "a few"), that should cut down on this type of support.