Add an initial Development Environment
This commit adds an initial Docker Compose development environment.
Not all services are catered for yet - this simple sets up a Postgres
database and an application server, there's no queue daemon or stubbing
of things like MapIt, DaDem or Ratty.

It should, however, be enough to make simple changes to the site's HTML
and CSS and can be extended in future.

Note that this imposes a constraint on `conf/packages` in that it
expects it to be simply a list of packages, one per line, without
version constraints.

We could use a package file with version constraints (which this one
doesn't currently have) but this would require the additional step of
converting that into the dependencies of a stub Debian package, then
installing that.