Add an initial Development Environment
This commit adds an initial Docker Compose development environment.
Not all services are catered for yet - this simple sets up a Postgres
database and an application server, there's no queue daemon or stubbing
of things like MapIt, DaDem or Ratty.

It should. however, be enough to make simple changes to the site's HTML
and CSS and can be extended in future.