Dockerfile: use repo packages file for installation
This reduces duplication, but requires the packages file to be simply a
list of packages, one per line, without version constraints.

We could use a package file with version constraints (which this one
doesn't currently have) but this would require the additional step of
converting that into the dependencies of a stub Debian package, then
installing that.

If needed, that could be added separately.