Increase grid gutter to 32px, to avoid footer overflow bug
Some of the footer elements create a -16px margin, to horizontally align
child elements. But in combination with WriteToThem’s 15px gutters, this
was causing a 1px horizontal scroll overflow on narrow screens.

Bumping the Foundation gutter setting up from 30px to 32px fixes it, and
the 2px difference in grid spacing is effectively unnoticeable.