Explicitly check for missing array keys
Newer versions of PHP are stricter about handling null values, so we
should test that a given key is present in an array rather than just
expecting a missing key not to match a value.

Otherwise from PHP7.4 we'll see `Trying to access array offset on value
of type null` errors. There should be no harm in fixing this now.