Use do-not-reply email for forwarding bounces.
We have started having an issue where an auto-reply from Parliament
to our VERP address, rather than the Reply-To, being forwarded to a
Microsoft email address, is being rejected. It passes both DKIM and
DMARC but fails SPF. Hopefully switching to an actual email for the
sender that we pass SPF on will deal with this.