translate who page strings
Note that some of the strings for this page are in votingarea.php which
is where xgettext finds them. However, they end up getting passed
through gettext again here because they are held in data structures that
get parsed before the locale is set. That's why there are things like
`$v['rep_name'] = _($va_rep_name[$vat]);` in the code.