store message language when creating the message
Use the LANGUAGE constant to pass the inteface language to the queue and
then store that in the message table.

Rather that write a whole new language handling thing for templates and
links just use the existing cobrand one which will do the same thing.
There is an assumption here that we're never going to do a multi
language cobrand but as we don't really do cobrands this seems like a
reasonable assumption to make in favour of less code