dummy DaDem endpoint to make development easier
This adds a dummy DaDem, plus GAZE and Ratty endpoint that fakes things
well enough to allow local testing without access to those services.
It's still a bit fragile but in general it works well enough when run
with Docker to allow you to do the whole thing.

Currently you have to log in to the container and start fyrqd by hand
for things to work.

The results on the candiated choice screen are somewhat unrealistic but
the point is more to let you run through the screens.

The actual script is outside the web tree and mapped in by Docker so
won't be accidentally discoverable live.